Create Bubble Sort
In the past, I always said that the sorting algorithm is simple, so I never want to do it. Of course, it is also because I am too lazy, dragging and dragging, and finally completely forgot. In fact, different sorting algorithms still have their own advantages. The sorting algorithm implemented today is simple in logic, but inefficient. The later, the more traversal time the sorting wastes.